+++ requirements.txt
@@ -1,2 +1,2 @@
-Flask==2.3.2
-# Jinja2==3.1.2 Shouldn't be necessary as Flask includes Jinja2
+Flask==2.1.0
+google-cloud-datastore==2.15.1
