+++ static/script.js
@@ -1,0 +1,7 @@
+'use strict';
+
+window.addEventListener('load', function () {
+
+  console.log("Hello World!");
+
+});
